Michael Josten: Changes: - changed the navigation drawer layout to have all the items that correlate to the core fragments of the chat app - changed the package name from PhishApp to Chapp - Added several Temporary Icons in Drawable for the Naviagtion Items - created drawable nav_drawer_item_bg -> item background that is supposed to be applied to all items in the navigation drawer, makes them look fancier. I'm not sure that it is working correctly. -In HomeActivity starting at line 169 -> added placement for handling onClicks for the navigation items, currently prints to console what item was pressed.  * Logout requires a seperate onClick listener because it is not a menu item, but is a TextView to have at the bottom of the drawer.  -In nav_header_home -> commented out several different varieties of navigation drawer styles, such as the navigation drawer on the right side and the original navigation of the PhishApp.
