+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "tcss450.uw.edu.phishapp"
+        applicationId "tcss450.uw.edu.chapp"
         minSdkVersion 24
         targetSdkVersion 28
         versionCode 1
@@ -32,4 +32,5 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    implementation 'com.android.support:mediarouter-v7:28.0.0'
 }
