+++ app/src/main/res/values/strings.xml
@@ -8,16 +8,23 @@
     <string name="text_phish_app">Chapp</string>
     <string name="text_password_confirm">Retype Password</string>
     <string name="title_activity_home">HomeActivity</string>
+    <string name="action_settings">Settings</string>
+    <string name="text_full_post">Full Post</string>
+    <string name="sample_text">sample</string>
+    <string name="nav_set_lists_text">Set Lists</string>
+
+    <!-- Naviation Drawer Strings -->
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
     <string name="nav_header_title">Chapp</string>
     <string name="nav_header_subtitle">The Best Chat App!</string>
     <string name="nav_header_desc">Navigation header</string>
-    <string name="action_settings">Settings</string>
     <string name="nav_home_text">Home</string>
     <string name="nav_blog_posts_text">Blog Posts</string>
-    <string name="text_full_post">Full Post</string>
-    <string name="sample_text">sample</string>
-    <string name="nav_set_lists_text">Set Lists</string>
+
+    <string name="nav_connections">Connections</string>
+    <string name="nav_chat">Chat</string>
+    <string name="nav_weather">Weather</string>
+    <string name="nav_logout">Logout</string>
 
 </resources>
